rename: Avatar web-domains common 으로 이동
