+++ packages/web-domains/src/common/components/Avatar/Avatar.tsx
