+++ packages/web-domains/src/result/features/components/SelectedDataCard/Profile.tsx
@@ -3,7 +3,7 @@
 import { Txt, TxtProps } from '@sambad/sds/components';
 import { HTMLAttributes } from 'react';
 
-import { Avatar } from '@/home/common/components/Avatar/Avatar';
+import { Avatar } from '@/common/components/Avatar/Avatar';
 
 import { profileCss } from './styles';
